Move ExternalPrivKey, ExternalPubKey to zcash_primitives.
