+++ zcash_primitives/Cargo.toml
@@ -28,6 +28,7 @@ equihash = { version = "0.1", path = "../components/equihash" }
 ff = "0.11"
 fpe = "0.5"
 group = "0.11"
+hdwallet = { version = "0.3.0", optional = true }
 hex = "0.4"
 incrementalmerkletree = "0.2"
 jubjub = "0.8"
