+++ zcash_primitives/src/lib.rs
@@ -17,6 +17,7 @@ pub mod memo;
 pub mod merkle_tree;
 pub mod sapling;
 pub mod transaction;
+pub mod transparent;
 pub mod zip32;
 pub mod zip339;
